requirements.txt
Flask
python-dotenv
openai>=1.0.0
